refactor(console): prepare API menu migration to angular for V1&V2 APIs
